+++ data/FarstriderLibData~Init.lua
@@ -1,9 +1,9 @@
 -- FarstriderLibData~Init.lua
 -- local _, FarstriderLibData = ...
 
-local VERSION = 10002
+local VERSION = 10100
 
-if _G.FarstriderLibData and (_G.FarstriderLibData.VERSION or 0) >= VERSION then return end
+if FarstriderLibData_API and (FarstriderLibData_API.VERSION or 0) >= VERSION then return end
 
 FarstriderLibData.VERSION = VERSION
 FarstriderLibData.Internal = {}